aliyun-fc3/code/requirements.txt
pyodps
pandas
numpy
tiktoken
langchain
langchain_openai
pydantic-core
langchain_community
loguru